+++ src/material.c
@@ -1,6 +1,6 @@
 #include "material.h"
 
-Material *new_material(double surfCol[NR_VEC_ELEMENTS], double refl, float shininess)
+Material *new_material(float surfCol[NR_VEC_ELEMENTS], float refl, float shininess)
 {
     Material *mat = (Material *)malloc(sizeof(Material));
     mat->surfCol[0] = surfCol[0];
